+++ src/ddzy/utility/algorithm/sort/insert-sort/index.ts
@@ -1,0 +1,15 @@
+/**
+ * 插入排序
+ * @param arr 源数组
+ */
+export default function insertSort(arr: number[]): number[] {
+  for (let i = 1; i < arr.length; i++) {
+    for (let j = i; j >= 0; j--) {
+      if (arr[j] < arr[j - 1]) {
+        [arr[j], arr[j - 1]] = [arr[j - 1], arr[j]];
+      }
+    }
+  }
+
+  return arr;
+}